rest-api/function-source-MNIST-OBJECT-pred/requirements.txt
tensorflow==2.10.0
google-cloud-storage==1.16.1
Pillow
flask==2.3.3